Renaming stomp_manager module
