+++ src/app/beer_garden/api/stomp/manager.py
